requirements.txt
#This file is a list of all pip requirements for your skill (if needed)
adapt-parser
gensim
mycroft-core